Added confirm windows for when the user is about to delete a site or …

+++ src/Components/Sites.js
@@ -37,8 +37,11 @@ var Sites = React.createClass({
 
   //deletes the selectedSite
   deleteSelectedSite() {
-    Api.deleteSite(this.state.selectedRowId).then(() => {window.location.reload()});
-
+    var confirmed = confirm('Are you sure you want to delete this site?');
+    if (confirmed) {
+      Api.deleteSite(this.state.selectedRowId).then(() => {window.location.reload()});
+    }
+    return;
   },
 
   goToAddSite() {
